request: adds set_cookie, reworks headers as a result
The original implementation of headers assumed that each response header
could only appear once, and should it appear multiple times, a
comma-seperated list could be used instead. While this is supposedly
standard, most browsers don't adhear to it, especially for 'Set-Cookie'.

As a result, headers were reimplemented to allow for duplicates. This
does increase complexity by quite a lot, but most of it can be ignored
by assuming each header only has one value and using 0 for the indicies.

Changes the URL encoding methods in strutil to take a talloc context to
make memory mangement easier. Modified test cases to account for this.

Rewrites a ton of header test cases and adds test cases for set_cookie.